When we push an encoded frame's data to disk because we can't write it yet, the idea was to set encoded in the QueueItem to 0 to say that the data has already been pushed to disk.
This was not working because we were resetting encoded in a copy of the QueueItem,
and hence not affecting the QueueItem held in _queue.

This meant that the same frame could be pushed to disk several times.

To fix this, clear encoded (with the lock held) once we have definitely decided
to push the frame to disk.